reverted entirely

+++ util.py
@@ -373,7 +373,7 @@ class ContrastiveLearningViewGenerator(object):
             transforms.Normalize((0.5, 0.5, 0.5), (0.5, 0.5, 0.5))])
         # print("My length is", len([self.base_transform(x) for i in range(self.n_views)] + [transform(x)]))
         # return [self.base_transform(x) for i in range(self.n_views)]
-        return self.base_transform(x), self.base_transform(x), transform(x) # dataloader handles the rest
+        return self.base_transform(x).double(), self.base_transform(x).double(), transform(x).double() # dataloader handles the rest
 
 
 class GaussianBlur(object):
